jd-research-tool/requirements.txt
opencv-python
pillow
pytesseract
python-dotenv
google-genai
